pyproject.toml
[tool.poetry]
name = "sideshow"
version = "0.1.0"
description = ""
packages = [
    { include = "sideshow" },
]
authors = ["Leeward Bound <leeward@boundcorp.net>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
mountaineer = "0.5.0"
django = "^5.0.0"
django-environ = "*"
django-cors-headers = "*"
django-rest-framework = "*"
django-mountaineer = { git = "https://github.com/boundcorp/django-mountaineer.git", tag = "v1.0.3" }
requests = "*"
django-extensions = "*"
django-filter = "*"
psycopg2-binary = "*"
uwsgi = "*"
whitenoise = "*"
django-minio-storage = "*"
bpython = "*"
pydantic = "*"
djantic2 = "*"
pytz = "*"
factory-boy = "*"
ruff = "*"
pytest = "*"
pytest-django = "*"  # Added pytest-django dependency
python-multipart = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.flake8]
ignore = ["E203", "E266", "E501", "W503", "F403", "F401"]
max-line-length = 88
max-complexity = 18
select = ["B", "C", "E", "F", "W", "T4"]

[tool.mypy]
files = ["sideshow"]
ignore_missing_imports = true
plugins = ["mypy_django_plugin.main", "pydantic.mypy"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "sideshow.settings.test_settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
addopts = "--ds=sideshow.settings.test_settings"